requirements.txt
cycler==0.10.0
kiwisolver==1.2.0
matplotlib==3.2.2
numpy==1.19.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
requests~=2.24.0
matplotlib-label-lines==0.3.8
PyYAML~=5.3.1